+++ NEWS.md
@@ -1,5 +1,7 @@
 # unpivotr (development version)
 
+* Make compatible with dplyr v1.1.0 (#58 @DavisVaughan)
+* Implement `as_cells.matrix()` (#52 @billdenney)
 * Allow behead() to work on grouped data (#46 @billdenney)
 * Add as_cells.matrix (#41 @billdenney)